Fix missing copyright headers
